requirements for PTT.txt
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.2
google-cloud-storage